fix get_count method and add lastInsertId property
get_count method was returning 1 no matter the query. By adjusting the method to account for the returned values, we get the true count.

The last inserted ID was available as a return variable on the insert method, but to make it a bit easier to get to, I added it as the lastInsertId property.